add new stacks landing page
