+++ assets/src/components/stacks/Stacks.tsx
@@ -2,122 +2,176 @@ import {
   EmptyState,
   Input,
   LoopingLogo,
-  PlusIcon,
   SearchIcon,
-  Sidecar,
-  SidecarItem,
-  SubTab,
-  TabList,
-  TreeNavEntry,
+  TabPanel,
+  Table,
+  useSetBreadcrumbs,
 } from '@pluralsh/design-system'
 import { useTheme } from 'styled-components'
-import React, { useEffect, useMemo, useRef, useState } from 'react'
+import React, {
+  ComponentProps,
+  Suspense,
+  useMemo,
+  useRef,
+  useState,
+} from 'react'
 import { isEmpty } from 'lodash'
-import { Outlet, useMatch, useNavigate, useParams } from 'react-router-dom'
-import capitalize from 'lodash/capitalize'
-
+import { useNavigate } from 'react-router-dom'
 import { useDebounce } from '@react-hooks-library/core'
-
-import Fuse from 'fuse.js'
-
-import moment from 'moment'
+import { type Row, createColumnHelper } from '@tanstack/react-table'
 
 import {
-  STACK_CONFIG_REL_PATH,
-  STACK_ENV_REL_PATH,
-  STACK_JOB_REL_PATH,
-  STACK_REPO_REL_PATH,
-  STACK_RUNS_REL_PATH,
+  STACKS_ABS_PATH,
   getStacksAbsPath,
 } from '../../routes/stacksRoutesConsts'
 import { GqlError } from '../utils/Alert'
-import { extendConnection, mapExistingNodes } from '../../utils/graphql'
-import { StackedText } from '../utils/table/StackedText'
+import { mapExistingNodes } from '../../utils/graphql'
 import { StackFragment, useStacksQuery } from '../../generated/graphql'
-import { RESPONSIVE_LAYOUT_CONTENT_WIDTH } from '../utils/layout/ResponsiveLayoutContentContainer'
-import { ResponsiveLayoutSidecarContainer } from '../utils/layout/ResponsiveLayoutSidecarContainer'
-import { ResponsiveLayoutSpacer } from '../utils/layout/ResponsiveLayoutSpacer'
+import { PluralErrorBoundary } from '../cd/PluralErrorBoundary'
+import LoadingIndicator from '../utils/LoadingIndicator'
+import { ResponsivePageFullWidth } from '../utils/layout/ResponsivePageFullWidth'
+import { FullHeightTableWrap } from '../utils/layout/FullHeightTableWrap'
+import { useFetchPaginatedData } from '../cd/utils/useFetchPaginatedData'
+import { Title1H1 } from '../utils/typography/Text'
+import { ColWithIcon } from '../utils/table/ColWithIcon'
+import { StackedText } from '../utils/table/StackedText'
 import { ClusterProviderIcon } from '../utils/Provider'
-import { ResponsiveLayoutPage } from '../utils/layout/ResponsiveLayoutPage'
-import { ResponsiveLayoutSidenavContainer } from '../utils/layout/ResponsiveLayoutSidenavContainer'
+import { TRUNCATE_LEFT } from '../utils/truncate'
 
-import { StandardScroller } from '../utils/SmoothScroller'
-
-import { LinkTabWrap } from '../utils/Tabs'
-
-import { StackTypeIcon, StackTypeIconFrame } from './StackTypeIcon'
 import CreateStack from './create/CreateStack'
-import StackDelete from './StackDelete'
-import StackStatusChip from './StackStatusChip'
+import { StackTypeIcon } from './common/StackTypeIcon'
+import StackStatusChip from './common/StackStatusChip'
+import StackApprovalChip from './common/StackApprovalChip'
 
-const pollInterval = 10 * 1000
+export const BREADCRUMBS = [{ label: 'stacks', url: STACKS_ABS_PATH }]
 
-const searchOptions = {
-  keys: ['name'],
-  threshold: 0.25,
+const QUERY_PAGE_SIZE = 100
+
+const REACT_VIRTUAL_OPTIONS: ComponentProps<
+  typeof Table
+>['reactVirtualOptions'] = {
+  overscan: 10,
 }
 
-const directory = [
-  { path: STACK_RUNS_REL_PATH, label: 'Runs' },
-  { path: STACK_CONFIG_REL_PATH, label: 'Configuration' },
-  { path: STACK_REPO_REL_PATH, label: 'Repository' },
-  { path: STACK_ENV_REL_PATH, label: 'Environment' },
-  { path: STACK_JOB_REL_PATH, label: 'Job' },
-] as const
+const COLUMN_HELPER = createColumnHelper<StackFragment>()
 
-export type StackOutletContextT = {
-  stack: StackFragment
-  refetch?: Nullable<() => void>
-}
+const COLUMNS = [
+  COLUMN_HELPER.accessor((stack) => stack, {
+    id: 'name',
+    header: 'Stack name / Repository', // TODO
+    meta: { truncate: true },
+    cell: ({ getValue }) => {
+      const stack = getValue()
 
-export const getBreadcrumbs = (stackId: string) => [
-  { label: 'stacks', url: getStacksAbsPath('') },
-  ...(stackId ? [{ label: stackId, url: getStacksAbsPath(stackId) }] : []),
+      return (
+        <ColWithIcon
+          truncateLeft
+          icon={
+            <StackTypeIcon
+              size={16}
+              stackType={stack.type}
+              css={{ filter: 'grayscale(1) brightness(2.5)' }} // TODO
+            />
+          }
+        >
+          {/* TODO: Change text props. */}
+          <StackedText
+            css={{ ...TRUNCATE_LEFT }}
+            first={stack.name}
+            second={stack.repository?.url}
+          />
+        </ColWithIcon>
+      )
+    },
+  }),
+  COLUMN_HELPER.accessor((stack) => stack, {
+    id: 'status',
+    header: 'Status',
+    cell: ({ getValue }) => {
+      const stack = getValue()
+
+      return (
+        <StackStatusChip
+          paused={!!stack.paused}
+          deleting={!!stack.deletedAt}
+        />
+      )
+    },
+  }),
+  COLUMN_HELPER.accessor((stack) => stack.git, {
+    id: 'git',
+    header: 'Folder / Ref', // TODO
+    cell: ({ getValue }) => {
+      const git = getValue()
+
+      return (
+        <StackedText
+          first={git.folder}
+          second={git.ref}
+        />
+      )
+    },
+  }),
+  COLUMN_HELPER.accessor((stack) => stack.cluster, {
+    id: 'cluster',
+    header: 'Cluster',
+    cell: ({ getValue }) => {
+      const cluster = getValue()
+
+      return (
+        <ColWithIcon
+          truncateLeft
+          icon={
+            <ClusterProviderIcon
+              cluster={cluster}
+              size={16}
+            />
+          }
+        >
+          {cluster?.name}
+        </ColWithIcon>
+      )
+    },
+  }),
+  COLUMN_HELPER.accessor((stack) => stack.approval, {
+    id: 'approval',
+    header: 'Approval',
+    cell: ({ getValue }) => <StackApprovalChip approval={!!getValue()} />,
+  }),
 ]
 
 export default function Stacks() {
   const theme = useTheme()
   const navigate = useNavigate()
-  const { stackId = '' } = useParams()
   const tabStateRef = useRef<any>(null)
-  const pathMatch = useMatch(`${getStacksAbsPath(stackId)}/:tab`)
-  const tab = pathMatch?.params?.tab || ''
-  const currentTab = directory.find(({ path }) => path === tab)
-  const [listRef, setListRef] = useState<any>(null)
   const [searchString, setSearchString] = useState('')
   const debouncedSearchString = useDebounce(searchString, 100)
 
-  const { data, error, loading, fetchMore, refetch } = useStacksQuery({
-    fetchPolicy: 'cache-and-network',
-    pollInterval,
-  })
+  useSetBreadcrumbs(BREADCRUMBS)
 
-  const { stacks, pageInfo } = useMemo(
-    () => ({
-      stacks: mapExistingNodes(data?.infrastructureStacks),
-      pageInfo: data?.infrastructureStacks?.pageInfo,
-    }),
+  const {
+    data,
+    loading,
+    error,
+    refetch,
+    pageInfo,
+    fetchNextPage,
+    setVirtualSlice,
+  } = useFetchPaginatedData(
+    {
+      queryHook: useStacksQuery,
+      pageSize: QUERY_PAGE_SIZE,
+      queryKey: 'infrastructureStacks',
+    },
+    {
+      q: debouncedSearchString,
+    }
+  )
+
+  const stacks = useMemo(
+    () => mapExistingNodes(data?.infrastructureStacks),
     [data?.infrastructureStacks]
   )
-
-  const stack = useMemo(
-    () => stacks.find(({ id }) => id === stackId),
-    [stackId, stacks]
-  )
-
-  // TODO: Use server-side filtering once it will be available.
-  const filteredStacks = useMemo(() => {
-    if (!debouncedSearchString) {
-      return stacks || []
-    }
-    const fuse = new Fuse(stacks || [], searchOptions)
-
-    return fuse.search(debouncedSearchString).map((result) => result.item)
-  }, [debouncedSearchString, stacks])
-
-  useEffect(() => {
-    if (!isEmpty(stacks) && !stackId) navigate(getStacksAbsPath(stacks[0].id))
-  }, [stacks, stackId, navigate])
 
   if (error)
     return (
@@ -133,191 +187,77 @@ export default function Stacks() {
     return <LoopingLogo />
   }
 
-  if (isEmpty(stacks)) {
-    return (
-      <EmptyState message="Looks like you don't have any infrastructure stacks yet.">
-        <CreateStack refetch={refetch} />
-      </EmptyState>
-    )
-  }
-
-  if (!stack) {
-    return <LoopingLogo />
-  }
-
   return (
-    <ResponsiveLayoutPage css={{ paddingBottom: theme.spacing.large }}>
-      <ResponsiveLayoutSidenavContainer width={360}>
+    <ResponsivePageFullWidth
+      headingContent={
         <div
           css={{
             display: 'flex',
-            gap: theme.spacing.small,
-            marginBottom: theme.spacing.medium,
+            gap: theme.spacing.large,
+            flexGrow: 1,
+            width: '100%',
           }}
         >
+          <Title1H1 css={{ flexGrow: 1, margin: 0 }}>
+            Infrastructure stacks
+          </Title1H1>
           <Input
             flexGrow={1}
-            placeholder="Search"
+            placeholder="Search by name"
             startIcon={<SearchIcon />}
             value={searchString}
             onChange={(e) => {
               setSearchString?.(e.currentTarget.value)
             }}
           />
-          <CreateStack
-            buttonContent={<PlusIcon />}
-            buttonProps={{ secondary: true, height: 40 }}
-            refetch={refetch}
-          />
+          <CreateStack refetch={refetch} />
         </div>
-        <StandardScroller
-          listRef={listRef}
-          setListRef={setListRef}
-          items={filteredStacks}
-          loading={loading}
-          placeholder={() => (
-            <div css={{ height: 52, borderBottom: theme.borders.default }} />
-          )}
-          hasNextPage={pageInfo?.hasNextPage}
-          mapper={(stack) => (
-            <TreeNavEntry
-              key={stack.id ?? ''}
-              label={
-                <div
-                  css={{
-                    alignItems: 'center',
-                    display: 'flex',
-                    gap: theme.spacing.small,
-                  }}
-                >
-                  <StackTypeIconFrame
-                    size="small"
-                    stackType={stack.type}
-                  />
-                  <StackedText
-                    first={stack.name}
-                    second={stack.repository?.url}
-                  />
-                </div>
-              }
-              active={stack.id === stackId}
-              activeSecondary={false}
-              href={getStacksAbsPath(stack.id)}
-              desktop
-            />
-          )}
-          loadNextPage={() =>
-            pageInfo?.hasNextPage &&
-            fetchMore({
-              variables: { after: pageInfo?.endCursor },
-              updateQuery: (
-                prev,
-                { fetchMoreResult: { infrastructureStacks } }
-              ) =>
-                extendConnection(
-                  prev,
-                  infrastructureStacks,
-                  'infrastructureStacks'
-                ),
-            })
-          }
-          refreshKey={undefined}
-          setLoader={undefined}
-          handleScroll={undefined}
-        />
-      </ResponsiveLayoutSidenavContainer>
-      <ResponsiveLayoutSpacer />
-      <div css={{ width: RESPONSIVE_LAYOUT_CONTENT_WIDTH }}>
-        <TabList
-          scrollable
-          gap="xxsmall"
+      }
+    >
+      <PluralErrorBoundary>
+        <TabPanel
+          css={{ height: '100%' }}
           stateRef={tabStateRef}
-          stateProps={{
-            orientation: 'horizontal',
-            selectedKey: currentTab?.path,
-          }}
-          marginRight="medium"
-          paddingBottom="small"
         >
-          {directory.map(({ label, path }) => (
-            <LinkTabWrap
-              subTab
-              key={path}
-              textValue={label}
-              to={`${getStacksAbsPath(stackId)}/${path}`}
+          <Suspense fallback={<LoadingIndicator />}>
+            <div
+              css={{
+                display: 'flex',
+                flexDirection: 'column',
+                gap: theme.spacing.small,
+                height: '100%',
+              }}
             >
-              <SubTab
-                key={path}
-                textValue={label}
-              >
-                {label}
-              </SubTab>
-            </LinkTabWrap>
-          ))}
-        </TabList>
-        <Outlet context={{ stack, refetch } as StackOutletContextT} />
-      </div>
-      <ResponsiveLayoutSpacer />
-      <ResponsiveLayoutSidecarContainer
-        display="flex"
-        flexDirection="column"
-        gap="small"
-      >
-        {stack && (
-          <>
-            <StackDelete
-              stack={stack}
-              refetch={refetch}
-            />
-            <Sidecar heading="Stack">
-              <SidecarItem heading="Name">
-                <div css={{ display: 'flex', gap: theme.spacing.small }}>
-                  {stack.name}
-                </div>
-              </SidecarItem>
-              <SidecarItem heading="ID">{stack.id}</SidecarItem>
-              <SidecarItem heading="Created">
-                {moment(stack.insertedAt).fromNow()}
-              </SidecarItem>
-              {stack.deletedAt && (
-                <SidecarItem heading="Deleted">
-                  {moment(stack.deletedAt).fromNow()}
-                </SidecarItem>
+              {!isEmpty(stacks) ? (
+                <FullHeightTableWrap>
+                  <Table
+                    columns={COLUMNS}
+                    reactTableOptions={{ meta: { refetch } }}
+                    reactVirtualOptions={REACT_VIRTUAL_OPTIONS}
+                    data={stacks || []}
+                    virtualizeRows
+                    hasNextPage={pageInfo?.hasNextPage}
+                    fetchNextPage={fetchNextPage}
+                    isFetchingNextPage={loading}
+                    onVirtualSliceChange={setVirtualSlice}
+                    onRowClick={(_e, { original }: Row<StackFragment>) =>
+                      navigate(getStacksAbsPath(original.id))
+                    }
+                    css={{
+                      maxHeight: 'unset',
+                      height: '100%',
+                    }}
+                  />
+                </FullHeightTableWrap>
+              ) : (
+                <EmptyState message="Looks like you don't have any infrastructure stacks yet.">
+                  <CreateStack refetch={refetch} />
+                </EmptyState>
               )}
-              <SidecarItem heading="Status">
-                <StackStatusChip
-                  paused={!!stack.paused}
-                  deleting={!!stack.deletedAt}
-                />
-              </SidecarItem>
-              <SidecarItem heading="Approval">
-                {stack.approval ? 'Required' : 'Not required'}
-              </SidecarItem>
-              <SidecarItem heading="Type">
-                <div css={{ display: 'flex', gap: theme.spacing.xsmall }}>
-                  <StackTypeIcon
-                    size={16}
-                    stackType={stack.type}
-                  />
-                  {capitalize(stack.type)}
-                </div>
-              </SidecarItem>
-              <SidecarItem heading="Repository">
-                {stack.repository?.url}
-              </SidecarItem>
-              <SidecarItem heading="Cluster">
-                <div css={{ display: 'flex', gap: theme.spacing.xsmall }}>
-                  <ClusterProviderIcon
-                    cluster={stack.cluster}
-                    size={16}
-                  />
-                  {stack.cluster?.name}
-                </div>
-              </SidecarItem>
-            </Sidecar>
-          </>
-        )}
-      </ResponsiveLayoutSidecarContainer>
-    </ResponsiveLayoutPage>
+            </div>
+          </Suspense>
+        </TabPanel>
+      </PluralErrorBoundary>
+    </ResponsivePageFullWidth>
   )
 }
