+++ assets/src/components/stacks/common/StackApprovalChip.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Chip } from '@pluralsh/design-system'
+
+import { ChipProps } from '@pluralsh/design-system/dist/components/Chip'
+
+export default function StackApprovalChip({
+  approval,
+  ...props
+}: {
+  approval: boolean
+} & ChipProps) {
+  return approval ? (
+    <Chip {...props}>Required</Chip>
+  ) : (
+    <Chip {...props}>Not required</Chip>
+  )
+}
