+++ assets/src/components/stacks/common/StackStatusChip.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { CheckIcon, Chip, PauseIcon, Spinner } from '@pluralsh/design-system'
+import { Chip } from '@pluralsh/design-system'
 
 import { ChipProps } from '@pluralsh/design-system/dist/components/Chip'
 
@@ -13,27 +13,21 @@ export default function StackStatusChip({
 } & ChipProps) {
   return deleting ? (
     <Chip
-      size="small"
       severity="danger"
-      icon={<Spinner />}
       {...props}
     >
       Deleting
     </Chip>
   ) : paused ? (
     <Chip
-      size="small"
       severity="warning"
-      icon={<PauseIcon />}
       {...props}
     >
       Paused
     </Chip>
   ) : (
     <Chip
-      size="small"
-      severity="success"
-      icon={<CheckIcon />}
+      severity="info"
       {...props}
     >
       Active
