+++ assets/src/components/stacks/create/CreateStack.tsx
@@ -1,32 +1,20 @@
-import { Button, Tooltip } from '@pluralsh/design-system'
-import { ButtonProps } from 'honorable'
-import { ReactNode, useState } from 'react'
+import { Button } from '@pluralsh/design-system'
+import { useState } from 'react'
 
 import { ModalMountTransition } from '../../utils/ModalMountTransition'
 
 import CreateStackModal from './CreateStackModal'
 
 export default function CreateStack({
-  buttonContent = 'Create stack',
-  buttonProps,
   refetch,
 }: {
-  buttonProps?: ButtonProps
-  buttonContent?: string | ReactNode
   refetch?: Nullable<() => void>
 }) {
   const [open, setOpen] = useState(false)
 
   return (
     <>
-      <Tooltip label="Create stack">
-        <Button
-          {...buttonProps}
-          onClick={() => setOpen(true)}
-        >
-          {buttonContent}
-        </Button>
-      </Tooltip>
+      <Button onClick={() => setOpen(true)}>Create stack</Button>
       <ModalMountTransition open={open}>
         <CreateStackModal
           open={open}
