+++ assets/src/components/stacks/run/Sidenav.tsx
@@ -6,8 +6,8 @@ import { AppIcon, Tab, TabList } from '@pluralsh/design-system'
 import { ResponsiveLayoutSidenavContainer } from '../../utils/layout/ResponsiveLayoutSidenavContainer'
 import { StackedText } from '../../utils/table/StackedText'
 import { StackRun } from '../../../generated/graphql'
-import { StackTypeIcon } from '../StackTypeIcon'
-import { StackRunStatusChip } from '../StackRunStatusChip'
+import { StackTypeIcon } from '../common/StackTypeIcon'
+import { StackRunStatusChip } from '../common/StackRunStatusChip'
 
 const DIRECTORY = [
   { path: '', label: 'Progress' },
