+++ assets/src/routes/stacksRoutes.tsx
@@ -1,4 +1,4 @@
-import { Navigate, Route } from 'react-router-dom'
+import { Route } from 'react-router-dom'
 
 import Stacks from '../components/stacks/Stacks'
 import StackRunDetail from '../components/stacks/run/Route'
@@ -9,25 +9,11 @@ import StackRunPlan from '../components/stacks/run/plan/Plan'
 
 import StackRunState from '../components/stacks/run/state/State'
 
-import StackRuns from '../components/stacks/StackRuns'
-
-import StackEnvironment from '../components/stacks/environment/StackEnvironment'
-import StackRepository from '../components/stacks/repository/StackRepository'
-
-import StackConfiguration from '../components/stacks/configuration/StackConfiguration'
-
-import StackJob from '../components/stacks/job/StackJob'
-
 import {
   STACKS_ABS_PATH,
-  STACK_CONFIG_REL_PATH,
-  STACK_ENV_REL_PATH,
-  STACK_JOB_REL_PATH,
-  STACK_REPO_REL_PATH,
   STACK_RUNS_ABS_PATH,
   STACK_RUNS_OUTPUT_REL_PATH,
   STACK_RUNS_PLAN_REL_PATH,
-  STACK_RUNS_REL_PATH,
   STACK_RUNS_REPOSITORY_REL_PATH,
   STACK_RUNS_STATE_REL_PATH,
 } from './stacksRoutesConsts'
@@ -36,37 +22,7 @@ export const stacksRoutes = [
   <Route
     path={STACKS_ABS_PATH}
     element={<Stacks />}
-  >
-    <Route
-      index
-      element={
-        <Navigate
-          replace
-          to={STACK_RUNS_REL_PATH}
-        />
-      }
-    />
-    <Route
-      path={STACK_RUNS_REL_PATH}
-      element={<StackRuns />}
-    />
-    <Route
-      path={STACK_CONFIG_REL_PATH}
-      element={<StackConfiguration />}
-    />
-    <Route
-      path={STACK_REPO_REL_PATH}
-      element={<StackRepository />}
-    />
-    <Route
-      path={STACK_ENV_REL_PATH}
-      element={<StackEnvironment />}
-    />
-    <Route
-      path={STACK_JOB_REL_PATH}
-      element={<StackJob />}
-    />
-  </Route>,
+  />,
   <Route
     path={STACK_RUNS_ABS_PATH}
     element={<StackRunDetail />}
@@ -93,3 +49,33 @@ export const stacksRoutes = [
     />
   </Route>,
 ]
+
+// {/* <Route */}
+// {/*  index */}
+// {/*  element={ */}
+// {/*    <Navigate */}
+// {/*      replace */}
+// {/*      to={STACK_RUNS_REL_PATH} */}
+// {/*    /> */}
+// {/*  } */}
+// {/* /> */}
+// {/* <Route */}
+// {/*  path={STACK_RUNS_REL_PATH} */}
+// {/*  element={<StackRuns />} */}
+// {/* /> */}
+// {/* <Route */}
+// {/*  path={STACK_CONFIG_REL_PATH} */}
+// {/*  element={<StackConfiguration />} */}
+// {/* /> */}
+// {/* <Route */}
+// {/*  path={STACK_REPO_REL_PATH} */}
+// {/*  element={<StackRepository />} */}
+// {/* /> */}
+// {/* <Route */}
+// {/*  path={STACK_ENV_REL_PATH} */}
+// {/*  element={<StackEnvironment />} */}
+// {/* /> */}
+// {/* <Route */}
+// {/*  path={STACK_JOB_REL_PATH} */}
+// {/*  element={<StackJob />} */}
+// {/* /> */}
